Fix null backdrop removal issue

+++ backend/app/Services/Storage.php
@@ -56,11 +56,15 @@
      * @param $poster
      * @param $backdrop
      */
-    public function removeImages($poster, $backdrop)
+    public function removeImages(string | null $poster, string | null $backdrop): void
     {
-      LaravelStorage::delete($poster);
-      LaravelStorage::disk('subpage')->delete($poster);
-      LaravelStorage::disk('backdrop')->delete($backdrop);
+      if($poster) {
+        LaravelStorage::delete($poster);
+        LaravelStorage::disk('subpage')->delete($poster);
+      }
+      if($backdrop) {
+        LaravelStorage::disk('backdrop')->delete($backdrop);
+      }
     }
 
     /**
